moved hostedPlotState to extensions
